+++ tests/test_didl_lite.py
@@ -366,3 +366,24 @@ class TestDidlLite:
 
         items = didl_lite.from_xml_string(didl_string)
         assert len(items) == 1
+
+    def test_item_set_attributes(self):
+        didl_string = """
+<DIDL-Lite xmlns="urn:schemas-upnp-org:metadata-1-0/DIDL-Lite/"
+           xmlns:upnp="urn:schemas-upnp-org:metadata-1-0/upnp/"
+           xmlns:dc="http://purl.org/dc/elements/1.1/"
+           xmlns:sec="http://www.sec.co.kr/">
+    <item restricted="1">
+        <dc:title>Video Item Title</dc:title>
+        <upnp:class>object.item.videoItem</upnp:class>
+    </item>
+</DIDL-Lite>"""
+
+        items = didl_lite.from_xml_string(didl_string)
+        assert len(items) == 1
+
+        item = items[0]
+        assert item.title == 'Video Item Title'
+        assert hasattr(item, 'rating')
+        assert item.rating is None
+        assert len(item.resources) == 0